+++ cmd/frontend/graphqlbackend/search.go
@@ -288,6 +288,9 @@ func (r *searchResolver) resolveRepositories(ctx context.Context, effectiveRepoF
 
 	archivedStr, _ := r.query.StringValue(query.FieldArchived)
 	archived := parseYesNoOnly(archivedStr)
+	if archived == Invalid {
+		archived = No // archived defaults to No.
+	}
 
 	commitAfter, _ := r.query.StringValue(query.FieldRepoHasCommitAfter)